Fold the status-envelope context back into one fact
SLOP-7 asked to extend file.statusEnvelopes with the parent-node kind. The
implementation instead added a second file.statusEnvelopeContexts fact and
joined the two by line:column at rule evaluation time. That meant two parallel
representations of the same array literal, two AST passes with two copies of
nodeStartColumn, a cached context entry for every array literal in the file
rather than only the envelopes, and a join that fails silently to
"assigned-generic-status-envelope" if the two column computations ever diverge.

file.statusEnvelopes now carries kind alongside statusKey, statusValue,
payloadKeys, line and column, produced in the parent-connected pass that
already had the parent links. file.statusEnvelopeContexts is deleted, as are
PhpFacts::statusEnvelopes and its helpers, which the port had superseded.

Rule evidence is unchanged: kind=, status=, payload= in the same order, with
returned / json / assigned classification verified against all three shapes.
php.structure schema goes to 5 for the changed fact set.

Also restores the original line endings and import order in DefaultRegistry.php
so its diff is the two-line registration change rather than a whole-file
rewrite.

Board: SLOP-2, SLOP-6 and SLOP-7 are implemented here, so they move to Done and
archive instead of sitting in READY/BACKLOG while their code is merged.
